Created ProductList component Used ProductList in product listing and drops
